+++ src/delombok/lombok/delombok/DocCommentIntegrator.java
@@ -31,11 +31,13 @@ import java.util.regex.Pattern;
 
 import com.sun.tools.javac.parser.Tokens.Comment;
 import com.sun.tools.javac.tree.DocCommentTable;
+import com.sun.tools.javac.tree.EndPosTable;
 import com.sun.tools.javac.tree.JCTree;
 import com.sun.tools.javac.tree.JCTree.JCClassDecl;
 import com.sun.tools.javac.tree.JCTree.JCCompilationUnit;
 import com.sun.tools.javac.tree.JCTree.JCMethodDecl;
 import com.sun.tools.javac.tree.JCTree.JCVariableDecl;
+import com.sun.tools.javac.util.JCDiagnostic.DiagnosticPosition;
 import com.sun.tools.javac.tree.TreeScanner;
 
 import lombok.javac.CommentInfo;
@@ -132,12 +134,41 @@ public class DocCommentIntegrator {
 					return docCommentContent_;
 				}
 				
+				@Override public DiagnosticPosition getPos() {
+					return new DiagnosticPosition() {
+						public JCTree getTree() {
+							return node;
+						}
+						
+						public int getStartPosition() {
+							return pos;
+						}
+						
+						public int getPreferredPosition() {
+							return pos;
+						}
+						
+						@SuppressWarnings("unused") // We compile against very old versions of javac intentionally (to support old stuff), but this is the method for newer impls.
+						public int getEndPosition(EndPosTable endPosTable) {
+							int end = endPosTable == null ? 0 : endPosTable.getEndPos(node);
+							if (end > pos) return end;
+							return pos + docCommentContent_.length();
+						}
+						
+						public int getEndPosition(Map<JCTree, Integer> endPosTable) {
+							Integer end = endPosTable.get(node);
+							if (end != null && end.intValue() > pos) return end.intValue();
+							return pos + docCommentContent_.length();
+						}
+					};
+				}
+				
 				@Override public int getSourcePos(int index) {
 					return pos + index;
 				}
 				
 				@Override public CommentStyle getStyle() {
-					return CommentStyle.JAVADOC;
+					return (CommentStyle) Javac.getCommentStyle();
 				}
 				
 				@Override public boolean isDeprecated() {