+++ src/utils/lombok/javac/Javac.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2009-2019 The Project Lombok Authors.
+ * Copyright (C) 2009-2025 The Project Lombok Authors.
  * 
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -22,7 +22,7 @@
 package lombok.javac;
 
 import static lombok.javac.JavacTreeMaker.TreeTag.treeTag;
-import static lombok.javac.JavacTreeMaker.TypeTag.typeTag;
+import static lombok.javac.JavacTreeMaker.TypeTag.*;
 
 import java.lang.reflect.Field;
 import java.lang.reflect.InvocationTargetException;
@@ -43,6 +43,7 @@ import com.sun.tools.javac.code.Symtab;
 import com.sun.tools.javac.code.Type;
 import com.sun.tools.javac.main.JavaCompiler;
 import com.sun.tools.javac.parser.Tokens.Comment;
+import com.sun.tools.javac.parser.Tokens.Comment.CommentStyle;
 import com.sun.tools.javac.tree.DocCommentTable;
 import com.sun.tools.javac.tree.JCTree;
 import com.sun.tools.javac.tree.JCTree.JCAnnotation;
@@ -183,7 +184,8 @@ public class Javac {
 	public static final TypeTag CTC_NONE = typeTag("NONE");
 	public static final TypeTag CTC_BOT = typeTag("BOT");
 	public static final TypeTag CTC_ERROR = typeTag("ERROR");
-	public static final TypeTag CTC_UNKNOWN = typeTag("UNKNOWN");
+	public static final TypeTag CTC_UNKNOWN = typeTagPermissive("UNKNOWN"); // UNKNOWN has been removed in JDK24, hence, we need to look it up permissively (just make it `null` if it does not exist).
+	
 	public static final TypeTag CTC_UNDETVAR = typeTag("UNDETVAR");
 	public static final TypeTag CTC_CLASS = typeTag("CLASS");
 	
@@ -330,6 +332,10 @@ public class Javac {
 		}
 	}
 	
+	public static Object getCommentStyle() {
+		return JavadocOps_8.COMMENT_STYLE;
+	}
+	
 	private static class JavadocOps_8 {
 		static String getJavadoc(Object dc, JCTree node) {
 			DocCommentTable dct = (DocCommentTable) dc;
@@ -350,6 +356,15 @@ public class Javac {
 			dct.putComment(node, newCmt);
 		}
 		
+		private static final CommentStyle COMMENT_STYLE = getCommentStyle();
+		private static CommentStyle getCommentStyle() {
+			try {
+				return CommentStyle.valueOf("JAVADOC");
+			} catch (IllegalArgumentException e) {
+				return CommentStyle.valueOf("JAVADOC_BLOCK");
+			}
+		}
+		
 		private static Comment createJavadocComment(final String text, final JCTree field) {
 			return new Comment() {
 				@Override public String getText() {
@@ -361,11 +376,15 @@ public class Javac {
 				}
 				
 				@Override public CommentStyle getStyle() {
-					return CommentStyle.JAVADOC;
+					return COMMENT_STYLE;
 				}
 				
 				@Override public boolean isDeprecated() {
 					return text.contains("@deprecated") && field instanceof JCVariableDecl && isFieldDeprecated(field);
+				}
+
+				@Override public DiagnosticPosition getPos() {
+					return field;
 				}
 			};
 		}
@@ -402,7 +421,7 @@ public class Javac {
 			throw sneakyThrow(e.getCause());
 		}
 	}
-	
+
 	public static void storeEnd(JCTree tree, int pos, JCCompilationUnit top) {
 		try {
 			Object endPositions = JCCOMPILATIONUNIT_ENDPOSITIONS.get(top);